refactor: move typeorm unit tests to its folder
