+++ test/adapters/typeorm/options/pagination.spec.ts
@@ -1,6 +1,6 @@
-import { ConfigProfile } from '../../../src/profile/config-profile'
-import { ENABLED_PROFILE } from '../../../src/profile/defaults'
-import { TypeORMQueryBuilder } from '../../../src/typeorm/query-builder'
+import { ConfigProfile } from '../../../../src/profile/config-profile'
+import { ENABLED_PROFILE } from '../../../../src/profile/defaults'
+import { TypeORMQueryBuilder } from '../../../../src/typeorm/query-builder'
 
 describe('Test Profiles for <pagination>', () => {
   let qb: TypeORMQueryBuilder
